+++ capa/features/extractors/ts/extractor.py
@@ -1,0 +1,136 @@
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import List, Tuple, Union, Iterator
+from pathlib import Path
+
+import capa.features.extractors.script
+import capa.features.extractors.ts.file
+import capa.features.extractors.ts.engine
+import capa.features.extractors.ts.global_
+import capa.features.extractors.ts.function
+import capa.features.extractors.ts.autodetect
+from capa.exceptions import UnsupportedFormatError
+from capa.features.common import Namespace
+from capa.features.address import NO_ADDRESS, Address, AbsoluteVirtualAddress, FileOffsetRangeAddress
+from capa.features.extractors.script import LANG_TEM, LANG_HTML
+from capa.features.extractors.ts.tools import BaseNamespace
+from capa.features.extractors.ts.engine import TreeSitterHTMLEngine, TreeSitterTemplateEngine, TreeSitterExtractorEngine
+from capa.features.extractors.ts.function import PSEUDO_MAIN, TSFunctionInner
+from capa.features.extractors.base_extractor import (
+    Feature,
+    BBHandle,
+    InsnHandle,
+    SampleHashes,
+    FunctionHandle,
+    StaticFeatureExtractor,
+)
+
+
+class TreeSitterFeatureExtractor(StaticFeatureExtractor):
+    engines: List[TreeSitterExtractorEngine]
+    template_engine: TreeSitterTemplateEngine
+    language: str
+    path: Path
+
+    def __init__(self, path: Path):
+        super().__init__(
+            # Tree-Sitter currently does not yield hash information about the sample in its output
+            hashes=SampleHashes(md5="", sha1="", sha256="")
+        )
+        self.path = path
+        with self.path.open("rb") as f:
+            buf = f.read()
+
+        try:
+            self.language = capa.features.extractors.ts.autodetect.get_language(path)
+            self.template_engine = self.get_template_engine(buf)
+            self.engines = self.get_engines(buf)
+        except ValueError as e:
+            raise UnsupportedFormatError(e)
+
+    def get_template_engine(self, buf: bytes):
+        if self.language == LANG_TEM:
+            return TreeSitterTemplateEngine(buf)
+
+    def get_engines(self, buf: bytes) -> List[TreeSitterExtractorEngine]:
+        if self.language == LANG_TEM and self.template_engine:
+            return self.extract_code_from_template()
+        if self.language == LANG_HTML:
+            return self.extract_code_from_html(buf)
+        return [TreeSitterExtractorEngine(self.language, buf)]
+
+    def extract_code_from_template(self) -> List[TreeSitterExtractorEngine]:
+        engines = list(self.template_engine.get_parsed_code_sections())
+        for node in self.template_engine.get_content_sections():
+            section_buf = self.template_engine.get_byte_range(node)
+            engines.extend(self.extract_code_from_html(section_buf, self.template_engine.namespaces))
+        return engines
+
+    def extract_code_from_html(
+        self, buf: bytes, namespaces: set[BaseNamespace] | None = None
+    ) -> List[TreeSitterExtractorEngine]:
+        if namespaces is None:
+            namespaces = set()
+        return list(TreeSitterHTMLEngine(buf, namespaces).get_parsed_code_sections())
+
+    def get_base_address(self) -> Union[AbsoluteVirtualAddress, capa.features.address._NoAddress]:
+        return NO_ADDRESS
+
+    def extract_template_namespaces(self) -> Iterator[Tuple[Feature, Address]]:
+        for ns in self.template_engine.get_namespaces():
+            address = NO_ADDRESS if ns.node is None else FileOffsetRangeAddress(ns.node.start_byte, ns.node.end_byte)
+            yield Namespace(ns.name), address
+
+    def extract_global_features(self) -> Iterator[Tuple[Feature, Address]]:
+        for engine in self.engines:
+            yield from capa.features.extractors.script.extract_language(engine.language, engine.get_default_address())
+        yield from capa.features.extractors.ts.global_.extract_features()
+
+    def extract_file_features(self) -> Iterator[Tuple[Feature, Address]]:
+        if self.language == LANG_TEM:
+            yield from self.extract_template_namespaces()
+        for engine in self.engines:
+            yield from capa.features.extractors.ts.file.extract_features(engine)
+
+    def get_pseudo_main_function_inner(self, engine: TreeSitterExtractorEngine) -> TSFunctionInner:
+        return TSFunctionInner(engine.tree.root_node, PSEUDO_MAIN, engine)
+
+    def get_pseudo_main_function(self, engine: TreeSitterExtractorEngine) -> FunctionHandle:
+        return FunctionHandle(engine.get_default_address(), self.get_pseudo_main_function_inner(engine))
+
+    def get_functions(self) -> Iterator[FunctionHandle]:
+        for engine in self.engines:
+            yield self.get_pseudo_main_function(engine)
+            for node in engine.get_function_definitions():
+                name_node = engine.get_function_definition_name(node)
+                name = engine.get_str(name_node) if name_node is not None else ""
+                yield FunctionHandle(engine.get_address(node), TSFunctionInner(node, name, engine))
+
+    def extract_function_features(self, f: FunctionHandle) -> Iterator[Tuple[Feature, Address]]:
+        yield from capa.features.extractors.ts.function.extract_features(f, f.inner.engine)
+
+    def get_basic_blocks(self, f: FunctionHandle) -> Iterator[BBHandle]:
+        yield from []
+
+    def extract_basic_block_features(self, f: FunctionHandle, bb: BBHandle) -> Iterator[Tuple[Feature, Address]]:
+        yield from []
+
+    def get_instructions(self, f: FunctionHandle, bb: BBHandle) -> Iterator[InsnHandle]:
+        yield from []
+
+    def extract_insn_features(
+        self, f: FunctionHandle, bb: BBHandle, insn: InsnHandle
+    ) -> Iterator[Tuple[Feature, Address]]:
+        yield from []