+++ capa/rules/__init__.py
@@ -183,6 +183,7 @@ SUPPORTED_FEATURES: dict[str, set] = {
         capa.features.common.OS,
         capa.features.common.Arch,
         capa.features.common.Format,
+        capa.features.common.ScriptLanguage,
     },
     Scope.FILE: {
         capa.features.common.MatchedRule,
@@ -446,6 +447,8 @@ def parse_feature(key: str):
         return capa.features.common.Namespace
     elif key == "property":
         return capa.features.insn.Property
+    elif key == "language":
+        return capa.features.common.ScriptLanguage
     else:
         raise InvalidRule(f"unexpected statement: {key}")